+++ src/aleph/__init__.py
@@ -4,7 +4,7 @@ import subprocess
 from pkg_resources import get_distribution, DistributionNotFound
 
 
-def get_git_version() -> str:
+def _get_git_version() -> str:
     output = subprocess.check_output(("git", "describe", "--tags"))
     return output.decode().strip()
 
@@ -14,6 +14,6 @@ try:
     dist_name = __name__
     __version__ = get_distribution(dist_name).version
 except DistributionNotFound:
-    __version__ = get_git_version()
+    __version__ = _get_git_version()
 finally:
     del get_distribution, DistributionNotFound
