+++ tests/api/test_version.py
@@ -1,7 +1,7 @@
 import pytest
 
+from aleph import __version__
 from aleph.web import app
-from aleph import get_git_version
 
 
 @pytest.mark.asyncio
@@ -12,4 +12,4 @@ async def test_get_version(aiohttp_client):
     assert response.status == 200, await response.text()
 
     data = await response.json()
-    assert data["version"] == get_git_version()
+    assert data["version"] == __version__
